Patch r81, r82, and r85 into the Beta branch to fix the remaining issues with unload handlers during shutdown.
r81: Fix not closing the browser with hung, crashed and interstitial tabs. Adds a 1 second hang monitor for the beforeunload/unload events to fire.

r82: Fix hang detection when closing a tab to not fire when a modal dialog is open.

r85: Fix broken unittest. AttemptToClosePage now sets up state. Do the same in the overridden method in TestRenderViewHost.


BUG=1296059
TBR=ojan



git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@88 4ff67af0-8c30-449e-8e8b-ad334ec8d88c