Merge r286 into Beta to fix bug 1303133.
This fixes http://b/issue?id=1303133, which was a crash
in the renderer due to the plugin instance getting destroyed
in the context of the ShowModalHTMLDialog request. The fix
is to convert the WebPluginDelegateProxy object in a DeleteLater and to set its channel to NULL in the destructor. This ensures that the reply to the message can be sent out successfully.

Bug=1303133
TBR= iyengar@google.com



git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@287 4ff67af0-8c30-449e-8e8b-ad334ec8d88c