Attempt to port r385 to the Beta branch to fix RTL titles in the very first dialog a user sees.
r385: Fixing an issue where BiDi text in window titles was not displayed correctly on Vista with aero.

BUG=1306227

NOTE: Due to the number of changes on trunk in window.cc, this is not a simple merge. I've updated the two sites where we are about to call SetWindowText() with the code to call AdjustStringForLocaleDirection().

R= idana@google.com

git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@419 4ff67af0-8c30-449e-8e8b-ad334ec8d88c