Rename HttpProxy* classes to Proxy*.  Move them into a net/proxy/ subdirectory.
I'm making this change because proxy resolution is really not specific to the HTTP protocol.  We need to use the proxy service in our FTP implementation, for example.  I made a separate directory instead of just putting these in base, because I anticipate more files once we have our own PAC implementation.

git-svn-id: http://src.chromium.org/svn/trunk/src@651 4ff67af0-8c30-449e-8e8b-ad334ec8d88c