Attempt at fixing flaky Find test.
The waterfall (build.chromium.org) has been complaining about this test for a while now. Nicolas and I looked at this, but we don't see anything wrong. We theorized this is because we issue Find right after the navigation and maybe the tab isn't ready. We'll see what happens with this change.

Review URL: http://codereview.chromium.org/2978

git-svn-id: http://src.chromium.org/svn/trunk/src@2411 4ff67af0-8c30-449e-8e8b-ad334ec8d88c