Experimental integration of delta compression content encoding
The command line option "sdch-enable" enables support of sdch
and automtic lazy download of dictionaries.  Optionally it
can select a singular domain to work from.

By default, all domains are enabled when the flag is used.
"-sdch-enable=".google.com"  Enables it only for Google.

When the switch is not set on the command line, all this
code is completely disabled.

Still TBD:
a) Finish implementation of security details (much of it is in place)
b) Add tests for security details.


r=huanr,ajenjo,kmixter

Review URL: http://codereview.chromium.org/461

git-svn-id: http://src.chromium.org/svn/trunk/src@2443 4ff67af0-8c30-449e-8e8b-ad334ec8d88c