Merge r2764 and r2774 to the release branch:
r3010 for sorting the URLs to open at startup depends on this.

Adds gfx::DisplayURL, which maintains a string as to how a URL
should be displayed along with a function to compare two DisplayURLs
for visually ordering them. This will be used both for displaying URLs
in a table as well as sorting them.

BUG=2949
TEST=none

TBR= sky@google.com
Review URL: http://codereview.chromium.org/7058

git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@3186 4ff67af0-8c30-449e-8e8b-ad334ec8d88c