Merge r3082 to the release branch:
Optimize the mime sniffer based on UMA data and move closer to the HTML 5 spec.

Removed over 50% of the heuristics while returning the same result 99.996% of
the time.  Each heuristic we remove reduces the attack surface that honest sites
have to worry about when the serve third-party content.

(with typo fix this time)

TBR= abarth@chromium.org

Review URL: http://codereview.chromium.org/7065

git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@3198 4ff67af0-8c30-449e-8e8b-ad334ec8d88c