Merge r3117, r3176, and r3178 to the release branch:
r3117: Changing the About box according to mockups. Basically, we now want to 
have a paragraph of text with embedded links (to the Chromium project, credits
and terms of service).

r3176: Allow the translators to chose which link appears first in the About 
box.

r3178: Forgot a .h file.

TBR= finnur@google.com

Review URL: http://codereview.chromium.org/7072

git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@3209 4ff67af0-8c30-449e-8e8b-ad334ec8d88c