Merge r3155 to the release branch:
Add a bunch of dlls to our dll eviction blacklist

Using the data collected by Marc-Antoine from 3800 dumps, I have selected the dlls that we
want to vote out of the island. This should reduce the number of crashes we suffer because
these dlls interact badly with the sandbox.
 - This change only afects renderers
 - We won't block it if the dll is not first found loaded in the browser process
 - Testing this is going to be challenging.

TBR= cpu@google.com
Review URL: http://codereview.chromium.org/7093

git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@3238 4ff67af0-8c30-449e-8e8b-ad334ec8d88c