Merge r3262 to the release branch:
This fixes http://code.google.com/p/chromium/issues/detail?id=2846, which is 
an issue with bloomberg.com not detecting the windows media player plugin. The
site reads the list of plugins off the Navigator object and runs through the 
list of plugins comparing each with the desired plugin name (Windows Media). 
This fails in Chrome as the Activex shim is registered by default as the 
handler for media player, with the name as activex-shim. 

The fix is to split up the Activex shim registration into two, one handling the
application/x-oleobject|application/oleobject mime types, and the other 
handling the windows media player types.

Bug=2846
TBR= ananta@chromium.org
Review URL: http://codereview.chromium.org/7258

git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@3276 4ff67af0-8c30-449e-8e8b-ad334ec8d88c