Merge r3228 to the release branch:
** 
Note: I changed line 548 in chrome\browser\download\download_manager.cc to
change the prefix from dangerous_download_ to unconfirmed_. I have a patch for 
review (http://codereview.chromium.org/7107) to make the same change on trunk.
**

This CL adds prompting for dangerous types of files (executable) when they are 
automatically downloaded.

The file is saved with a temporary name (dangerous_download_xxxx.download) in 
the download directory and the user is presented (in the download shelf and 
the download tab if opened) with a warning message and buttons to save/discard 
the download.

If discarded the download is removed (and its file deleted).
If saved, download goes as usual.

Dangerous downloads not confirmed by the user are deleted on shutdown.

TEST=Download a small exe file, try using the save/discard button from the 
download shelf and from the download tab (the intent is that the file has been 
entirely downloaded by the time you take action). Try again with a slow/big 
download (that time the download is expected not to be finished when 
approved/discarded).

TBR= jcampan@chromium.org
Review URL: http://codereview.chromium.org/7108

git-svn-id: http://src.chromium.org/svn/branches/chrome_official_branch/src@3285 4ff67af0-8c30-449e-8e8b-ad334ec8d88c