A fix for Issue 2673 in chromium: "Flash: IME not available".
To investigate this issue deeply, I noticed this issue consisted of two issues:

1. Windows cannot attach IMEs to a plug-in process.
Somehow, Windows attaches IMEs (i.e. loads IME DLLs, attaches them to a process, and adds message hooks to a window) while creating a top-level window. On the other hand, a plug-in process does not create any top-level windows. These facts prevent Windows from attaching IMEs to windowed plug-ins.

2. The CallWindowProc() function does not dispatch messages to the message-hook functions used by IMEs.
When the |plugin_wnd_proc_| value is a handle referring the DefWindowProc() function, it seems the CallWindowProc() does not dispatch messages to the hook functions used by IMEs.

To fix the former issue, this change once creates a top-level window and have Windows attach IMEs to this window, and change the window to a client window of the given parent window.
To fix the latter issue, this change creates a dummy window procedure WebPluginDelegateImpl::DummyWindowProc(), which just encapsulates the DefWindowProc(), and use it instead of the DefWindowProc() function.

BUG=2673
Review URL: http://codereview.chromium.org/4325
Review URL: http://codereview.chromium.org/4325

git-svn-id: http://src.chromium.org/svn/trunk/src@3391 4ff67af0-8c30-449e-8e8b-ad334ec8d88c