Add a lot of robustness to SDCH filter decoding
a) Do meta-refresh (and partially disable sdch) when we have an sdch decode
error early on (due to lack of dictionary).

b) Gracefully handle proxy removal of "gzip" in "content-encoding: sdch,gzip". 

c) Recover from completely stripped content encoding (in response to a
request for which we advertised an SDCH dictionary) by using a filter
that can optionally decode sdch with an optional gzip as well.

Some field tests have shown problems with this protocol's 
transit across the ether. The general approach we've adopted
is to stop using SDCH on the problematic site when we have problems. This
will prevent a malicious site from shutting down all use of SDCH.

r=kmixter,openvcdiff,huanr

Review URL: http://codereview.chromium.org/6433

git-svn-id: http://src.chromium.org/svn/trunk/src@3420 4ff67af0-8c30-449e-8e8b-ad334ec8d88c