Fix a problem with rdi.ca where the video (silverlight) is not painting.
This bug was fixed in 1735, but the code to fix that
got reverted (by mistake?) in revision 2744.

Comment for 1735:
The basic issue here was that the plugin would not paint correctly.
The URLs mentioned in this bug load windowed silverlight plugin instances,
which invoke the NPN_InvalidateRect API to paint. We send over the rects
to the renderer, however these don't generate paints as the plugin is
windowed. A peek at Safari's webkit implementation revealed that they
merely invoke the InvalidateRect windows API in this context.

The fix is to emulate this behavior.


bug:666
Review URL: http://codereview.chromium.org/7370

git-svn-id: http://src.chromium.org/svn/trunk/src@3481 4ff67af0-8c30-449e-8e8b-ad334ec8d88c